Improve loading indication. Add Replacement feature. Change main feature to contains. Change result pane to not disappear when no value.

+++ src/components/TestCases.js
@@ -1,34 +1,69 @@
 import {useState} from "react";
 import TestCaseItem from "./TestCaseItem";
+import ButtonSelector from "./ButtonSelector";
+import ReplacementTestCaseItem from "./ReplacementTestCaseItem";
 
-const renderTestcases = (testcases, updateItem, deleteItem) => {
+const renderTestcases = (testcases, mode, updateItem, deleteItem) => {
     let arr = [];
 
     testcases.forEach(elem => {
-        arr.push(
-            <TestCaseItem
-                key={elem.id}
-                id={elem.id}
-                value={elem.value}
-                pass={elem.pass}
-                onChange={(id, newValue) => {
-                    updateItem(id, newValue);
-                }}
-                deleteAction={(id) => {
-                    deleteItem(id);
-                }}
-            />
-        );
+        if(elem.mode === mode) {
+            if (elem.mode === "contains") {
+                arr.push(
+                    <TestCaseItem
+                        key={elem.id}
+                        id={elem.id}
+                        value={elem.value}
+                        pass={elem.pass}
+                        onChange={(id, newValue) => {
+                            updateItem(id, newValue);
+                        }}
+                        deleteAction={(id) => {
+                            deleteItem(id);
+                        }}
+                    />
+                );
+            } else if (elem.mode === "replacement") {
+                arr.push(
+                    <ReplacementTestCaseItem
+                        key={elem.id}
+                        id={elem.id}
+                        value={elem.value}
+                        result={elem.result}
+                        onChange={(id, newValue) => {
+                            updateItem(id, newValue);
+                        }}
+                        deleteAction={(id) => {
+                            deleteItem(id);
+                        }}
+                    />
+                );
+            }
+        }
     });
 
     return arr;
 };
 
 const TestCases = ({ value }) => {
-    const [testcases, setTestcases] = useState([{ id: 0, value: "test1234" }]);
+    const [testcases, setTestcases] = useState([]);
+    const [replacementValue, setReplacementValue] = useState("");
+    const [mode, setMode] = useState("contains");
 
-    const deleteitem = (id) => {
-        setTestcases(testcases.filter(x => x.id !== id));
+    const addItem = () => {
+        let new_tcases = testcases.concat([{
+            id: testcases.length,
+            value: "",
+            mode: mode
+        }]);
+        new_tcases.forEach((x, i) => x.id = i);
+        setTestcases(new_tcases);
+    };
+
+    const deleteItem = (id) => {
+        let new_arr = testcases.filter(x => x.id !== id);
+        new_arr.forEach((x, i) => x.id = i);
+        setTestcases(new_arr);
     };
 
     const updateItem = (id, newValue) => {
@@ -41,51 +76,80 @@ const TestCases = ({ value }) => {
         setTestcases(arr_copy);
     };
 
+    const getTestcaseCount = () => {
+        let count = 0;
+        testcases.forEach(elem => count += (elem.mode === mode) ? 1 : 0);
+        return count;
+    };
+
     const runTestCases = () => {
-        let exp = new RegExp(value);
+        let exp = new RegExp(value, "gi");
 
         let testcases_copy = testcases.filter(x => true);
 
-        for(let i = 0; i < testcases_copy.length; i++)
-            testcases_copy[i].pass = exp.test(testcases_copy[i].value);
+        for(let i = 0; i < testcases_copy.length; i++) {
+            if (testcases_copy[i].mode === "contains" && testcases_copy[i].mode === mode)
+                testcases_copy[i].pass = exp.test(testcases_copy[i].value);
+            else if (testcases_copy[i].mode === "replacement" && testcases_copy[i].mode === mode)
+                testcases_copy[i].result = testcases_copy[i].value.replaceAll(exp, replacementValue);
+        }
 
-        console.log(testcases_copy);
         setTestcases(testcases_copy);
     };
 
     return (
         <div className={"content-panel primary-color"}>
-            <label for={"tcases"} className={"form-label"}>Test Cases</label>
-            <div id={"tcases"}>
-                {
-                    (testcases.length > 0) ?
-                        renderTestcases(testcases, updateItem, deleteitem) :
-                        <div>
-                            <p>No test cases. Add one!</p>
-                        </div>
-                }
+            <div>
+                <ButtonSelector
+                    mode={mode}
+                    setMode={setMode}
+                />
             </div>
             <div>
-                <div className={"d-flex justify-content-between"}>
-                    <div>
-                        <button
-                            className={"btn btn-primary"}
-                            onClick={() => {
-                                let new_tcases = testcases.concat([{
-                                    id: testcases.length,
-                                    value: ""
-                                }]);
-                                setTestcases(new_tcases);
-                            }}
-                        >Add</button>
-                    </div>
-                    <div>
-                        <button
-                            className={"btn btn-secondary"}
-                            onClick={() => {
-                                runTestCases();
-                            }}
-                        >Run Test Cases</button>
+                <div>
+                    {
+                        (mode === "replacement") ?
+                            <div className={"mb-3"}>
+                                <label htmlFor={"replacement-value"} className={"form-label"}>Replacement Value</label>
+                                <div id={"replacement-value"}>
+                                    <input
+                                        type={"text"}
+                                        className={"form-control"}
+                                        value={replacementValue}
+                                        onChange={event => setReplacementValue(event.target.value)}
+                                    />
+                                </div>
+                            </div> : <div></div>
+                    }
+                </div>
+                <label htmlFor={"tcases"} className={"form-label"}>Test Cases</label>
+                <div id={"tcases"}>
+                    {
+                        (getTestcaseCount() > 0) ?
+                            renderTestcases(testcases, mode, updateItem, deleteItem) :
+                            <div>
+                                <p>No test cases. Add one!</p>
+                            </div>
+                    }
+                </div>
+                <div>
+                    <div className={"d-flex justify-content-between"}>
+                        <div>
+                            <button
+                                className={"btn btn-primary"}
+                                onClick={() => {
+                                    addItem();
+                                }}
+                            >Add</button>
+                        </div>
+                        <div>
+                            <button
+                                className={"btn btn-secondary"}
+                                onClick={() => {
+                                    runTestCases();
+                                }}
+                            >Run Test Cases</button>
+                        </div>
                     </div>
                 </div>
             </div>
